(PC-38491)[API] fix: SQL function get_deposit_balance prevents from spending last cents
(cherry picked from commit bfee5a7be1639780e98b439e0038db9a0874c648)

+++ api/src/pcapi/core/bookings/models.py
@@ -521,7 +521,7 @@ Booking.trig_ddl = f"""
         RETURNS numeric
         AS $$
     DECLARE
-        deposit_amount bigint := (SELECT CASE WHEN "expirationDate" > now() THEN amount ELSE 0 END amount FROM deposit WHERE id = deposit_id);
+        deposit_amount numeric := (SELECT CASE WHEN "expirationDate" > now() THEN amount ELSE 0 END amount FROM deposit WHERE id = deposit_id);
         sum_bookings numeric;
     BEGIN
         IF deposit_amount IS NULL
